FixnumArray experiment for reduced IR size
This experiment was intended to reduce the size of large numeric
arrays in memory in IR, especially those that are executed only
once to initialize global state. Specifically, the "mail" gem was
seen to have many extremely large arrays of integers as part of
its generated parsers (from Ragel). These arrays are used for
state transitions in the parser, but are only created once. The
large Operand[] held in the IR remains in memory forever (megabytes
of data in this case).

The logic was to store only the primitive values, and lazily create
a cached object array for future construction of Array objects. To
avoid holding a cached array for arrays allocated once and
abandoned, I only cached starting with the second execution, and
subsequent array construction created dup'ed instances (with
shared data) from then on.

Initially it was hoped that most of these arrays would be in the
range of unsigned bytes and a ByteArray operand would be enough,
but few fit of them fit that requirement. Expanding that to a
LongArray or FixnumArray worked as expected, but without carrying
over the cached Fixnum operands from the parser, duplicate values
ended up creating separate RubyFixnum instances and consuming far
more memory than the Operand[] the parser created originally.

Ultimately, I could not figure out a way to meet all of these
criteria:

* Compact storage of the array IR with only necessary data (and
  ideally smallest primitive array, though I did not attempt that).
* Efficient construction of first and n + 1th array objects without
  pinning too much data in memory.
* Efficient use of cached fixnum IR and objects to avoid
  constructing duplicate instances.

This could be revisited at some point, and also underscores the
importance of having the parser cache Fixnum instances.

The lazy caching + duping behavior could be ported into the
existing Array operand when its values are known to be all
immutable or immediate values, reducing greatly the cost to
repeatedly construct literal arrays that won't be modified.